Add DeletePost
Currently just accepts a URL and only returns an error if there's an
issue within url.Parse or http.Get. Mirroring Pinboard APIs lack of
concern if an item is requested for deletion but doesn't exist. (The
entire API is pretty lenient and returns status via XML blobs which is a
pain to parse, so avoiding it where it's not needed)